update styling, add react state
